Move types from CustomerSignUp to types file
